Разделил презентеры страниц (фильмы отдельно, статистика отдельно)
